Save trait for Dawgs
