+++ src/lib.rs
@@ -22,3 +22,4 @@ pub mod stat_utils;
 pub mod tokenize;
 pub mod weight;
 pub mod memory_backing;
+pub mod io;
